Add EncoderLineBreakOutFeed utility class
